# openapi.yaml
openapi: 3.0.3
info:
  title: AI Holiday Planner API
  description: An intelligent holiday planning API powered by Cerebras LLM.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
tags:
  - name: Health
    description: Health check endpoints
  - name: Holiday Planner
    description: Holiday planning and itinerary management endpoints
paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Verify that the API is running.
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/v1/planner/models:
    get:
      tags:
        - Holiday Planner
      summary: Get available models
      description: Retrieve a list of available model providers and their models.
      operationId: getModels
      responses:
        '200':
          description: List of models retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/planner/generate:
    post:
      tags:
        - Holiday Planner
      summary: Generate initial itinerary
      description: Create a personalized holiday itinerary based on user description.
      operationId: generateItinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryRequest'
      responses:
        '200':
          description: Itinerary generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/planner/refine:
    post:
      tags:
        - Holiday Planner
      summary: Refine existing itinerary
      description: Refine an itinerary based on user feedback while maintaining conversation context.
      operationId: refineItinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefinementRequest'
      responses:
        '200':
          description: Itinerary refined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryResponse'
        '400':
          description: Invalid request or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/planner/save:
    post:
      tags:
        - Holiday Planner
      summary: Save itinerary to file
      description: Save the current itinerary to a text file.
      operationId: saveItinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveItineraryRequest'
      responses:
        '200':
          description: Itinerary saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveItineraryResponse'
        '400':
          description: Invalid request or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/planner/config/model:
    post:
      tags:
        - Holiday Planner
      summary: Update default model configuration
      description: Update the default AI provider and model in the backend .env file.
      operationId: updateModelConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Configuration updated successfully
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/planner/sessions/{session_id}:
    get:
      tags:
        - Holiday Planner
      summary: Get session state
      description: Retrieve the current itinerary and follow-up questions for a session.
      operationId: getSession
      parameters:
        - name: session_id
          in: path
          description: Unique session identifier (UUID).
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryResponse'
        '400':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Holiday Planner
      summary: Delete session
      description: Delete a session and all its data.
      operationId: deleteSession
      parameters:
        - name: session_id
          in: path
          description: Unique session identifier (UUID).
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: healthy
        message:
          type: string
          example: API is running successfully
    ModelItem:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: llama-3.3-70b
        name:
          type: string
          example: Llama 3.3 70B
    ProviderInfo:
      type: object
      required:
        - provider
        - name
        - models
      properties:
        provider:
          type: string
          example: cerebras
        name:
          type: string
          example: Cerebras
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelItem'
    ModelsResponse:
      type: object
      required:
        - providers
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderInfo'
    ItineraryRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 10
          maxLength: 1000
          example: I want a 7-day tropical vacation in Bali with a moderate budget.
        provider:
          type: string
          default: cerebras
          example: cerebras
        model:
          type: string
          default: llama-3.3-70b
          example: llama-3.3-70b
    RefinementRequest:
      type: object
      required:
        - session_id
        - feedback
      properties:
        session_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        feedback:
          type: string
          minLength: 5
          maxLength: 1000
          example: I prefer more adventure activities.
    FollowUpQuestion:
      type: object
      required:
        - question
        - order
      properties:
        question:
          type: string
          example: Would you prefer beachfront or jungle accommodations?
        order:
          type: integer
          example: 1
    ItineraryResponse:
      type: object
      required:
        - session_id
        - itinerary
        - follow_up_questions
        - provider
        - model
      properties:
        session_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        itinerary:
          type: string
          example: "# 7-Day Bali Itinerary..."
        follow_up_questions:
          type: array
          items:
            $ref: '#/components/schemas/FollowUpQuestion'
        provider:
          type: string
          example: cerebras
        model:
          type: string
          example: llama-3.3-70b
    SaveItineraryRequest:
      type: object
      required:
        - session_id
      properties:
        session_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        filename:
          type: string
          maxLength: 100
          example: bali_trip
    SaveItineraryResponse:
      type: object
      required:
        - success
        - filename
        - message
      properties:
        success:
          type: boolean
          example: true
        filename:
          type: string
          example: itinerary_bali_trip.txt
        message:
          type: string
          example: Itinerary is being saved to itinerary_bali_trip.txt
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: API Key Missing
        detail:
          type: string
          example: CEREBRAS_API_KEY not found in environment variables
    ConfigUpdateRequest:
      type: object
      required:
        - provider
        - model
      properties:
        provider:
          type: string
          example: cerebras
        model:
          type: string
          example: llama-3.3-70b